-- make column::searchSorted virtual so that HDF5_FastQuery can specialize them, address warnings issued by gcc -Wall (--> ibis1.0.6)
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@126 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/ifade.cpp
@@ -265,7 +265,7 @@ int ibis::fade::read(const char* f) {
     // nbases, cnts, and bases
     uint32_t nb;
     ierr = UnixSeek(fdes, end, SEEK_SET);
-    if (ierr != end) {
+    if (ierr != (int)end) {
 	LOGGER(ibis::gVerbose > 0)
 	    << "ibis::fade::read(" << fnm << ") failed to seek to " << end;
 	UnixClose(fdes);
